Move projects update

+++ heimbohrtechnik/heim_bohrtechnik/page/bohrplaner/bohrplaner.js
@@ -754,7 +754,7 @@ frappe.bohrplaner = {
                 },
                 {
                     'fieldname': 'days', 
-                    'fieldtype': 'Int', 'label': __("Days"), 'default': 1
+                    'fieldtype': 'Float', 'label': __("Days"), 'default': 1
                 }
             ],
             'primary_action': function(){
